requirements.txt
pathway
llm_app
unstructured[pdf]
tiktoken
pandas
requests
datetime
streamlit
python-dotenv